consolidate and move type/interface definitions to their relevant modules in the component folder for cohesion and organization
