+++ client/src/components/AdditionalInfo/index.ts
@@ -1,0 +1,5 @@
+import { AdditionalInfoForm } from './AdditionalInfoForm';
+import { AdditionalInfo, AdditionalInfoComment } from './additionalInfoSchema';
+
+export { AdditionalInfoForm };
+export type { AdditionalInfo, AdditionalInfoComment };
