+++ client/src/components/HaztrakSite/index.ts
@@ -1,0 +1,5 @@
+import { HtSiteTable } from './SiteTable';
+import { HaztrakSite } from './haztrakSiteSchema';
+
+export { HtSiteTable };
+export type { HaztrakSite };
