--- client/src/components/Manifest/AdditionalInfo/index.ts
@@ -1,3 +1,0 @@
-import { AdditionalInfoForm } from './AdditionalInfoForm';
-
-export { AdditionalInfoForm };
