+++ client/src/components/Manifest/Handler/index.ts
@@ -2,5 +2,7 @@ import { HandlerForm } from './HandlerForm';
 import { HandlerSearchForm } from './HandlerSearchForm';
 import { TransporterSearchForm } from './TransporterSearchForm';
 import { AddHandler } from './AddHandler';
+import { Handler, handlerSchema, Signer } from './handlerSchema';
 
-export { HandlerForm, TransporterSearchForm, HandlerSearchForm, AddHandler };
+export { HandlerForm, TransporterSearchForm, HandlerSearchForm, AddHandler, handlerSchema };
+export type { Handler, Signer };
