+++ client/src/components/Manifest/Transporter/TransporterTable/TransporterTable.spec.tsx
@@ -1,9 +1,9 @@
 import { screen } from '@testing-library/react';
 import React from 'react';
 import { renderWithProviders } from 'test-utils';
-import { Transporter } from 'types/site';
 import { TransporterTable } from './index';
 import { createMockTransporter } from 'test-utils/fixtures';
+import { Transporter } from 'components/Manifest/Transporter/transporterSchema';
 
 const HANDLER_ID_1 = 'siteId1';
 const HANDLER_ID_2 = 'siteId2';
