+++ client/src/components/Manifest/Transporter/index.ts
@@ -1,4 +1,6 @@
 import { AddTransporter } from './AddTransporter';
 import { TransporterTable } from './TransporterTable';
+import { Transporter } from './transporterSchema';
 
 export { AddTransporter, TransporterTable };
+export type { Transporter };
