+++ client/src/components/Manifest/WasteLine/WasteLineForm.tsx
@@ -1,11 +1,11 @@
 import { HtCard, HtForm } from 'components/Ht';
-import { AdditionalInfoForm } from 'components/Manifest/AdditionalInfo';
+import { AdditionalInfoForm } from 'components/AdditionalInfo';
 import { HazardousWasteForm } from './HazardousWasteForm';
 import React from 'react';
 import { Button, Container, Form, Row } from 'react-bootstrap';
 import { FormProvider, UseFieldArrayAppend, useForm } from 'react-hook-form';
 import { Manifest } from 'components/Manifest/manifestSchema';
-import { WasteLine } from 'types/wasteLine';
+import { WasteLine } from 'components/Manifest/WasteLine/wasteLineSchema';
 import { QuantityForm } from './QuantityForm';
 
 interface WasteLineFormProps {
