+++ client/src/components/Manifest/WasteLine/index.ts
@@ -1,4 +1,6 @@
 import { AddWasteLine } from './AddWasteLine';
 import { WasteLineTable } from './WasteLineTable';
+import { WasteLine } from 'components/Manifest/WasteLine/wasteLineSchema';
 
 export { WasteLineTable, AddWasteLine };
+export type { WasteLine };
