+++ client/src/components/RcraSite/index.ts
@@ -1,0 +1,21 @@
+import { RcraSiteDetails } from './RcraSiteDetails';
+import { AddressListGroup } from './AddressListGroup';
+import {
+  RcraAddress,
+  rcraContactSchema,
+  RcraLocality,
+  rcraLocalitySchema,
+  rcraPhoneSchema,
+  rcraSite,
+  RcraSite,
+} from './rcraSiteSchema';
+
+export {
+  AddressListGroup,
+  RcraSiteDetails,
+  rcraSite,
+  rcraLocalitySchema,
+  rcraContactSchema,
+  rcraPhoneSchema,
+};
+export type { RcraSite, RcraAddress, RcraLocality };
