--- client/src/components/RcraSiteDetails/index.ts
@@ -1,4 +1,0 @@
-import { RcraSiteDetails } from './RcraSiteDetails';
-import { AddressListGroup } from './AddressListGroup';
-
-export { AddressListGroup, RcraSiteDetails };
