+++ client/src/features/haztrakSite/SiteDetails/SiteDetails.tsx
@@ -1,14 +1,18 @@
-import { RcraSiteDetails } from 'components/RcraSiteDetails';
+import { RcraSiteDetails } from 'components/RcraSite';
 import { HtCard } from 'components/Ht';
 import { useHtAPI } from 'hooks';
 import React, { ReactElement } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import { HaztrakSite } from 'types/site';
 import { SyncManifestBtn } from 'components/buttons';
 import { Container, Button } from 'react-bootstrap';
+import { HaztrakSite } from 'components/HaztrakSite';
 
 /**
  * GET and Display details of the Haztrak site including RCRA site details.
+ *
+ * This could be expanded to include other information about a haztrak site
+ * if, for example, the site also had information relevant to the office of water (OW)
+ * of data that's not directly related to any regulatory agency (the site's users).
  * @constructor
  */
 export function SiteDetails(): ReactElement {
