--- client/src/types/manifest/index.ts
@@ -1,17 +1,0 @@
-import { Manifest } from './manifest';
-import {
-  AdditionalInfo,
-  AdditionalInfoComment,
-} from 'components/Manifest/AdditionalInfo/additionalInfoSchema';
-import { ElectronicSignature, Signer } from './signatures';
-import { CorrectionInfo, CorrectionRequest } from './correction';
-
-export type {
-  Manifest,
-  AdditionalInfo,
-  AdditionalInfoComment,
-  CorrectionInfo,
-  CorrectionRequest,
-  ElectronicSignature,
-  Signer,
-};
