--- client/src/types/site/index.ts
@@ -1,18 +1,0 @@
-import { ElectronicSignature, Signer } from 'types/manifest';
-import { RcraSite, Handler, Transporter, HaztrakSite } from 'types/site/sites';
-
-import { RcraAddress, RcraLocality } from './address';
-import { RcraContact, RcraPhone } from './contact';
-
-export type {
-  HaztrakSite,
-  RcraSite,
-  Transporter,
-  Signer,
-  RcraAddress,
-  RcraLocality,
-  Handler,
-  ElectronicSignature,
-  RcraContact,
-  RcraPhone,
-};
